feat[nav]: nav feature add nav feature first push KA1-38
