+++ index.html
@@ -12,7 +12,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Do+Hyeon&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="./css/style.css" />
-    <script src="./js/footerHeader.js" defer></script>
+    <script type="module" src="./js/index.js" defer></script>
   </head>
   <body>
     <div class="tablet__container">
@@ -55,6 +55,27 @@
           </div>
         </div>
       </header>
+      <nav class="nav">
+        <section class="nav__sec">
+          <h2 class="nav__heading">배달</h2>
+          <ul class="nav__ul">
+            <li class="nav__li"><a href="#" class="selected">돈까스•회•일식</a></li>
+            <li class="nav__li"><a href="#">중식</a></li>
+            <li class="nav__li"><a href="#">치킨</a></li>
+            <li class="nav__li"><a href="#">백반•죽•국수</a></li>
+            <li class="nav__li"><a href="#">카페•디저트</a></li>
+            <li class="nav__li"><a href="#">분식</a></li>
+            <li class="nav__li"><a href="#">짬•탕•찌개</a></li>
+            <li class="nav__li"><a href="#">피자</a></li>
+            <li class="nav__li"><a href="#">양식</a></li>
+            <li class="nav__li"><a href="#">고기•구이</a></li>
+            <li class="nav__li"><a href="#">족발•보쌈</a></li>
+            <li class="nav__li"><a href="#">아시안</a></li>
+            <li class="nav__li"><a href="#">패스트푸드</a></li>
+            <li class="nav__li"><a href="#">도시락</a></li>
+          </ul>
+        </section>
+      </nav>
       <footer class="footer">
         <div class="footer__sub">
           <p class=""><a href="#">사업자정보확인 </a><span aria-hidden="true">|</span> <a href="#">이용약관 </a> <span aria-hidden="true">|</span> <a href="#">전자금융거래이용약관 </a>  <span aria-hidden="true">|</span> <a href="#" class="footer__bold">개인정보처리방침 </a></p>
