+++ js/index.js
@@ -1,0 +1,4 @@
+import footer from "./footer.js";
+import header from "./header.js";
+header();
+footer();
